Bug 571190 build the solution on Framework services host
Any application, that serves as Passage (licensing agent or Floating
server) can declare a Framework for it's own runtime to empower Passage
facilities.

From the other side, if this application itself is licensed with
Passage, it should have sort of `seal` bundle with another Framework
declaration.

This bundle's units are renamed according to these statements. This
bundle itself is to be renamed as well.

Signed-off-by: eparovyshnaya <elena.parovyshnaya@gmail.com>