Bug 575166 License Agreement management: license issuing
Roll back FeatureVersion.agreements reference addition.

Product-related agreements cannot be associated with FeatureVersion in
order to be implanted to an issued license: one of base Passage use
cases is to update a separate feature (install newer version, probably
with new set of Agreements) in scope of an old license (version range
allows). It is decided to protect Requirement gathering with
product-related agreements.

Signed-off-by: eparovyshnaya <elena.parovyshnaya@gmail.com>