Remove SLF4J Plugin from o.e.passage.lic.oshi feature
If a Feature includes a Plugin, it usually includes it with a specific
name and the version that was in the TP when the feature was build.
This prevents consumers from using a slf4j bundle with different
symbolic-name or different version in their TP or product.

This is part of
https://github.com/eclipse-platform/eclipse.platform.releng.aggregator/issues/588