Use Import-Package to reference junit
Don't use a version range for Test-Plugin dependencies because the
Target-Platform defines and controls a closed environment in which the
tests run. Because a version range is therefore not necessary for
Test-Plugins the version range can be omitted in order to simplify
version updates.

Fixes https://github.com/eclipse-passage/passage/issues/1224